Rename pdata/log to logs, move internal wrapper where all data are
